add users resolvers
