+++ api/package.json
@@ -2,6 +2,7 @@
   "dependencies": {
     "@graphql-tools/load-files": "^6.5.3",
     "@graphql-tools/schema": "^8.3.6",
+    "@types/lodash": "^4.14.181",
     "@types/socket.io": "^3.0.2",
     "aws-sdk": "^2.1018.0",
     "body-parser": "*",
@@ -10,6 +11,7 @@
     "express-graphql": "^0.12.0",
     "graphql": "^15.5.1",
     "graphql-tag": "^2.12.6",
+    "lodash": "^4.17.21",
     "pg": "^8.0.3",
     "pg-hstore": "^2.3.4",
     "reflect-metadata": "^0.1.13",
